+++ web/__tests__/business/hooks/indicator/use-history-indicators-value-view-model.hook.test.tsx
@@ -4,7 +4,7 @@ import { resetMockDB, mockDB } from '@/app/mocks/db';
 import { resetAllStore } from '@/app/store/stores/reset-store';
 import { useHistoryIndicatorsValueViewModel } from '@/app/business/hooks/indicator/use-history-indicators-value-view-model.hook';
 import { useWorkspaceStore } from '@/app/store/stores/numerical-guidance/workspace.store';
-import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
+import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 
 const wrapper = SWRProviderWithoutCache;
