+++ web/app/business/hooks/custom-forecast-indicator/use-custom-forecast-indicator-value-by-metadata.hook.ts
@@ -1,5 +1,5 @@
 import { useFetchCustomForecastIndicatorsValue } from '@/app/store/querys/numerical-guidance/custom-forecast-indicator.query';
-import { useSelectedIndicatorBoardMetadata } from '../indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
+import { useSelectedIndicatorBoardMetadata } from '../numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 
 export const useCustomForecastIndicatorsValueByMetadata = () => {
   const { selectedMetadata } = useSelectedIndicatorBoardMetadata();
