+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/editable-metadata-title.tsx
@@ -1,4 +1,4 @@
-import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
+import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
 import EditableText from '../../view/atom/editable-text';
 import clsx from 'clsx';
