+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-delete-dialog.tsx
@@ -2,7 +2,7 @@
 import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 import AlertDialog from '../../view/molecule/alert-dialog';
 import { IndicatorBoardMetadata } from '@/app/business/services/view-model/indicator-board-metadata-view-model.service';
-import { useIndicatorBoardMetadataList } from '@/app/business/hooks/indicator-board-metedata/use-indicator-board-metadata-list-view-model.hook';
+import { useIndicatorBoardMetadataList } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-list-view-model.hook';
 
 export default function MetadataDeleteDialog() {
   const { deleteIndicatorBoardMetadata } = useIndicatorBoardMetadataList();
