+++ web/app/ui/components/numerical-guidance/indicator-board-metadata/metadata-list/metadata-list.tsx
@@ -1,10 +1,10 @@
 'use client';
 import React from 'react';
 import List from '../../../view/molecule/list';
-import { useIndicatorBoardMetadataList } from '@/app/business/hooks/indicator-board-metedata/use-indicator-board-metadata-list-view-model.hook';
+import { useIndicatorBoardMetadataList } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-list-view-model.hook';
 import Button from '../../../view/atom/button/button';
 import Pending from '../../../view/molecule/pending';
-import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
+import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 import MetadataListItem from './metadata-list-item';
 import { IndicatorBoardMetadata } from '@/app/business/services/view-model/indicator-board-metadata-view-model.service';
 import { PlusIcon } from '@heroicons/react/solid';
