+++ web/app/ui/components/numerical-guidance/indicator/advanced-indicators-chart.tsx
@@ -4,7 +4,7 @@ import { useLiveIndicatorsValueViewModel } from '@/app/business/hooks/indicator/
 import { useEffect } from 'react';
 import { createIndicatorFormatter } from '@/app/business/services/chart/indicator-formatter.service';
 import { useCustomForecastIndicatorsValueViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-custom-forecast-indicators-value-view-model.hook';
-import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
+import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
 
 type AdvancedIndicatorsChartProps = {
   indicatorBoardMetadataId?: string;
